♻️ Split Projectinator's mixin-lookup responsibilities into MixinResolvinator
Projectinator had two unrelated jobs: loading the project file itself
(load, config_available?, lookup_yaml_extension) and resolving/
validating mixin references (extract_mixins, validate_mixin_load_paths,
validate_mixins, lookup_mixins). The mixin-lookup half now lives on its
own class, MixinResolvinator, wired into Composinator alongside the
now-leaner Projectinator.

This also drops two unused constructor dependencies Projectinator
never actually referenced in its own body: :system_wrapper (dead
regardless of this split) and :ruby_expandinator (moves with
extract_mixins, its only user).

Existing projectinator_mixin_spec.rb coverage moves to a new
mixin_resolvinator_spec.rb unchanged in substance -- this is a pure
structural move, not a behavior change.
